+++ client/src/reducers/userReducer.js
@@ -9,6 +9,7 @@ import { isEmptyString } from '../components/util/HelperFunctions.js';
 
 /* User */
 const userInitialState = {
+    uuid: '',
     firstName: '',
     lastName: '',
     avatar: '/favicon-96x96.png',
@@ -98,6 +99,7 @@ export default function userReducer(state = userInitialState, action) {
         case 'SET_USER_INFO':
             return {
                 ...state,
+                uuid: action.payload.uuid,
                 firstName: action.payload.firstName,
                 lastName: action.payload.lastName,
                 roles: action.payload.roles,