+++ server/api/users.js
@@ -19,6 +19,7 @@ const basicUserInfo = (req, res) => {
         uuid: req.decoded.uuid
     }, {
         _id: 0,
+        uuid: 1,
         firstName: 1,
         lastName: 1,
         roles: 1,